Safari "fix" removed
Safari doesn't reflow properly when resized; font size in particular is
not recalculated. I'd added a forced reflow on a resize event.

The problem on ipad/iphone is that when scrolling down, safari shrinks
the location bar, which grows the viewport, which in turn triggers the
resize event. So my 'fix' was causing blinking glitches while scrolling
down on safari.

Removed my fix, which will mean that the font-size may not change on
resize. May have to rely on media queries after all

	modified:   index.html