Lazy loading done correctly, based on onLoad events rather than arbitrary timing delays.
We want to optimize FCP (first content paint), and loading of several
scripts was blocking the main thread. Second, the scroll indicators
for the text boxes are effective only after the elements have been
laid out, which doesn't happen if the background image has not yet
loaded.

So we use the onload of the bottom image to trigger both.

	modified:   assets/js/lets.js
	modified:   index.html